changed the file structure to be coherent with rest of BE
