+++ controllers/edit/approveEdit.js
@@ -1,0 +1,15 @@
+const { models } = require('../../models');
+const { createError, GENERIC_ERROR } = require('../../helpers/error.js');
+
+/**
+ * @description Get Edits
+ *
+ * @param {object} req
+ * @param {object} res
+ * @param {object} next
+ */
+
+const getAllEdits = async (req, res, next) => {
+    try{
+        const approval = 
+    }
